Add support for Android SDK 1.0.1
