+++ sdk/android/build.gradle
@@ -2,15 +2,15 @@ group 'com.namiml.flutter.sdk'
 version '1.0-SNAPSHOT'
 
 buildscript {
-    ext.kotlin_version = '1.4.10'
-    ext.nami_sdk_version = '1.0.0'
+    ext.kotlin_version = '1.4.30'
+    ext.nami_sdk_version = '1.0.1'
     repositories {
         google()
         jcenter()
     }
 
     dependencies {
-        classpath 'com.android.tools.build:gradle:4.0.1'
+        classpath 'com.android.tools.build:gradle:4.1.2'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
     }
 }
